Create HumanPlayer class with get_move method
Created HumanPlayer class as part of making it possible to play games
with any combination of human/computer players. Players can now be
injected into the Game at the start, and the Game expects them to
implement a marker property and a get_move method. The get_move method consumes a board and a UI.