Change all Board() inits to use None not 'number'
Changed all Board() initialisations to use None values in the init
array for empty cells instead of using the cell number as a string.
This is in order to remove the last explicit reference to the player
as 'X' and 'O' from board and instead test that the value != None.