Add optional input function parameter to ConsoleUi
Add optional functional parameter to ConsoleUi to specify the function
to call to get user input. This is to allow the ConsoleUi to be tested
without replacing the global sys.stdin. The built in raw_input
function is injected by default, but during testing, a function which
sequentially returns test input can be injected.